feat - moved report functions to new class
