+++ src/Tickets/Flexible_Tickets/Series_Passes.php
@@ -16,6 +16,7 @@ use TEC\Events_Pro\Custom_Tables\V1\Templates\Series_Filters;
 use TEC\Tickets\Flexible_Tickets\Series_Passes\Labels;
 use TEC\Tickets\Flexible_Tickets\Series_Passes\Meta;
 use TEC\Tickets\Flexible_Tickets\Series_Passes\Queries;
+use TEC\Tickets\Flexible_Tickets\Series_Passes\Reports;
 use Tribe__Events__Main as TEC;
 use Tribe__Repository__Interface as ORM;
 use Tribe__Tickets__Ticket_Object as Ticket_Object;
@@ -83,6 +84,15 @@ class Series_Passes extends Controller {
 	private Queries $queries;
 
 	/**
+	 * A reference to the Series Passes' reports handler.
+	 *
+	 * @since TBD
+	 *
+	 * @var Reports
+	 */
+	private Reports $reports;
+
+	/**
 	 * Series_Passes constructor.
 	 *
 	 * since TBD
@@ -98,6 +108,7 @@ class Series_Passes extends Controller {
 		Labels $labels,
 		Meta $meta,
 		Metabox $metabox,
+		Reports $reports,
 		Ticket_Provider_Handler $ticket_provider_handler,
 		Queries $queries
 	) {
@@ -105,6 +116,7 @@ class Series_Passes extends Controller {
 		$this->labels                  = $labels;
 		$this->meta                    = $meta;
 		$this->metabox                 = $metabox;
+		$this->reports                 = $reports;
 		$this->ticket_provider_handler = $ticket_provider_handler;
 		$this->queries                 = $queries;
 	}
@@ -132,6 +144,8 @@ class Series_Passes extends Controller {
 	protected function do_register(): void {
 		$this->container->singleton( Series_Passes\Repository::class, Series_Passes\Repository::class );
 		$this->container->singleton( Series_Passes\Metadata::class, Series_Passes\Metadata::class );
+		// Register reporting functions.
+		$this->reports->register_hooks();
 
 		add_filter( 'the_content', [ $this, 'reorder_series_content' ], 0 );
 		add_action( 'tribe_events_tickets_new_ticket_buttons', [ $this, 'render_form_toggle' ] );
@@ -219,6 +233,8 @@ class Series_Passes extends Controller {
 	 * @return void
 	 */
 	public function unregister(): void {
+		// Unregister reporting functions.
+		$this->reports->unregister_hooks();
 		remove_filter( 'the_content', [ $this, 'reorder_series_content' ], 0 );
 		remove_action( 'tribe_events_tickets_new_ticket_buttons', [ $this, 'render_form_toggle' ] );
 		remove_action( 'admin_menu', [ $this, 'enable_reports' ], 20 );
